% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_comparison_data.R
\name{create_comparison_data}
\alias{create_comparison_data}
\title{Create Comparison Data}
\usage{
create_comparison_data(records, types, breaks, file_sizes, duplicates)
}
\arguments{
\item{file_sizes}{The size of each file}

\item{duplicates}{Indicators of whether each file possibly has duplicates}
}
\value{

}
\description{
Description goes here.
}
