% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs Sampler for Posterior Inference}
\usage{
gibbs_sampler(
  comparison_list,
  partition_prior_list,
  n_iter = 1000,
  mus = NA,
  nus = NA,
  Z_init = 1:r,
  seed = 70
)
}
\arguments{
\item{comparison_list}{the output from a call to
\code{create_comparison_data} or \code{reduce_comparison_data}}

\item{partition_prior_list}{the output of a call to
\code{specify_partition_prior}}

\item{n_iter}{The number of iterations of the Gibbs sampler to run}

\item{seed}{The seed to use while running the Gibbs sampler}
}
\value{
a list containing:
\describe{
  \item{\code{match_samp}}{}
  \item{\code{non_match_samp}}{}
  \item{\code{partitions}}{}
  \item{\code{contingency_tables}}{}
  \item{\code{cluster_sizes}}{}
  \item{\code{sampling_time}}{}
}
}
\description{
Description goes here.
}
