% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_comparison_data.R
\name{reduce_comparison_data}
\alias{reduce_comparison_data}
\title{Reduce Comparison Data Size}
\usage{
reduce_comparison_data(comparison_list, pairs_to_keep, cc = 1)
}
\arguments{
\item{comparison_list}{the output from a call to
\code{\link{create_comparison_data}}}

\item{pairs_to_keep}{a logical vector indicating which record pairs should be
kept}

\item{cc}{an indicator of whether we should use the connected components
induced by pairs_to_keep}
}
\value{
a list containing:
\describe{
  \item{\code{record_pairs}}{}
  \item{\code{comparisons}}{}
  \item{\code{K}}{}
  \item{\code{file_sizes}}{}
  \item{\code{duplicates}}{}
  \item{\code{field_levels}}{}
  \item{\code{file_labels}}{}
  \item{\code{rp_to_fp}}{}
  \item{\code{ab}}{}
  \item{\code{file_sizes_not_included}}{}
  \item{\code{ab_not_included}}{}
  \item{\code{labels}}{}
  \item{\code{pairs_to_keep}}{}
  \item{\code{cc}}{}
}
}
\description{
Description goes here.
}
\examples{

}
