% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_partition_prior.R
\name{specify_partition_prior}
\alias{specify_partition_prior}
\title{Specify the Partition Prior}
\usage{
specify_partition_prior(
  comparison_list,
  flat,
  alphas = NA,
  dup_upper_bound = NA,
  dup_count_prior_family = NA,
  dup_count_prior_pars = NA,
  n_prior_family = NA,
  n_prior_pars = NA
)
}
\arguments{
\item{comparison_list}{the output from a call to
\code{\link{create_comparison_data}} or \code{\link{reduce_comparison_data}}}

\item{n_prior_pars}{}
}
\value{
a list containing:
\describe{
  \item{\code{flat}}{}
  \item{\code{no_dups}}{}
  \item{\code{alphas}}{}
  \item{\code{alpha_0}}{}
  \item{\code{dup_upper_bound}}{}
  \item{\code{log_dup_count_prior}}{}
  \item{\code{log_n_prior}}{}
}
}
\description{
Description goes here.
}
