% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_partition.R
\name{initialize_partition}
\alias{initialize_partition}
\title{Initialize the Partition}
\usage{
initialize_partition(comparison_list, pairs_to_keep, seed = NA)
}
\arguments{
\item{comparison_list}{the output from a call to
\code{\link{create_comparison_data}} or \code{\link{reduce_comparison_data}}}

\item{seed}{}
}
\value{

}
\description{
Description goes here.
}
